Добавить новые провайдеры, исправить замечания к первой лабе

+++ rest/api/database/models/index.go
@@ -1,9 +1,9 @@
 package models
 
 type Index struct {
-	Id          int `json:"id"`
-	PastYear    int `json:"pastYear"`
-	CurrentYear int `json:"currentYear"`
-	Value       int `json:"value"`
-	CurrencyId  int `json:"currencyId"`
+	Id          int     `json:"id"`
+	PastYear    int     `json:"pastYear"`
+	CurrentYear int     `json:"currentYear"`
+	Value       float64 `json:"value"`
+	CurrencyId  int     `json:"currencyId"`
 }
